# docs/api/openapi.yaml
{"openapi":"3.1.0","info":{"title":"Plugin Gateway API","description":"First-party internal API for BC-02 Plugin Gateway REST proxying, webhook dispatch, and MCP routing for plugin tools.","version":"2026.07.1","summary":"Internal plugin and MCP gateway API","x-adapstory-api-major":"v1","x-adapstory-api-audience":"internal","x-adapstory-ai-ready":true},"servers":[{"url":"http://localhost","description":"Generated server url"}],"security":[{"bearerAuth":[]}],"paths":{"/internal/plugins/v1/{slug}/mcp":{"get":{"tags":["mcp-route-controller"],"operationId":"proxyMcpStream","parameters":[{"name":"slug","in":"path","required":true,"schema":{"type":"string"},"example":"example-slug"}],"responses":{"200":{"description":"OK","headers":{"X-Response-Id":{"description":"Technical response identifier equal to the inbound X-Request-Id.","schema":{"type":"string"}}},"content":{"text/event-stream":{"schema":{"$ref":"#/components/schemas/StreamingResponseBody"}}}}},"x-AI-ready":true},"post":{"tags":["mcp-route-controller"],"summary":"Proxy MCP JSON-RPC call to plugin backend","description":"Routes a JSON-RPC 2.0 MCP request to the plugin pod identified by slug. Injects X-Tenant-Id, X-Request-Id, X-Correlation-Id headers.","operationId":"proxyMcp","parameters":[{"name":"slug","in":"path","description":"Plugin slug identifier (e.g. 'course-builder')","required":true,"schema":{"type":"string","default":""},"example":"example-slug"}],"responses":{"200":{"description":"MCP response proxied from plugin backend","headers":{"X-Response-Id":{"description":"Technical response identifier equal to the inbound X-Request-Id.","schema":{"type":"string"}}}},"400":{"description":"Invalid plugin slug format","headers":{"X-Response-Id":{"description":"Technical response identifier equal to the inbound X-Request-Id.","schema":{"type":"string"}}}},"502":{"description":"Plugin pod unreachable or returned error","headers":{"X-Response-Id":{"description":"Technical response identifier equal to the inbound X-Request-Id.","schema":{"type":"string"}}}}},"x-AI-ready":true},"delete":{"tags":["mcp-route-controller"],"operationId":"terminateMcpSession","parameters":[{"name":"slug","in":"path","required":true,"schema":{"type":"string"},"example":"example-slug"}],"responses":{"200":{"description":"OK","headers":{"X-Response-Id":{"description":"Technical response identifier equal to the inbound X-Request-Id.","schema":{"type":"string"}}}}},"x-AI-ready":true}},"/internal/mcp-grants/v1":{"post":{"tags":["mcp-grant-controller"],"summary":"Register exact MCP provider bindings for an exchanged access token","description":"Revalidates the complete capability binding set through Plugin Lifecycle and stores one immutable, short-lived authorization record keyed by the token jti hash.","operationId":"register","parameters":[{"name":"X-Tenant-Id","in":"header","description":"Tenant claim; must exactly match the signed access token","required":true,"schema":{"type":"string","format":"uuid","default":"","example":"550e8400-e29b-41d4-a716-446655440000","maxLength":512,"minLength":36},"example":"550e8400-e29b-41d4-a716-446655440000"},{"name":"X-User-Id","in":"header","description":"Authenticated caller identity in service:<azp> form","required":true,"schema":{"type":"string","default":"","example":"service:adapstory-bc10-gateway-exchange","maxLength":512},"example":"service:plugin-lifecycle"},{"name":"X-Adapstory-User-Id","in":"header","description":"Delegated actor identity; must exactly match the exchanged token subject","required":true,"schema":{"type":"string","default":"","maxLength":512},"example":"teacher@example.com"}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/McpGrantRegistrationRequest"}}},"required":true},"responses":{"204":{"description":"Exact token-bound grant registered","headers":{"X-Response-Id":{"description":"Technical response identifier equal to the inbound X-Request-Id.","schema":{"type":"string"}}}},"401":{"description":"Gateway access token is invalid","headers":{"X-Response-Id":{"description":"Technical response identifier equal to the inbound X-Request-Id.","schema":{"type":"string"}}}},"403":{"description":"Signed and asserted identity differ","headers":{"X-Response-Id":{"description":"Technical response identifier equal to the inbound X-Request-Id.","schema":{"type":"string"}}}},"409":{"description":"Binding set is ambiguous or token is rebound","headers":{"X-Response-Id":{"description":"Technical response identifier equal to the inbound X-Request-Id.","schema":{"type":"string"}}}},"422":{"description":"Binding metadata is stale or invalid","headers":{"X-Response-Id":{"description":"Technical response identifier equal to the inbound X-Request-Id.","schema":{"type":"string"}}}},"503":{"description":"Lifecycle or Redis is unavailable","headers":{"X-Response-Id":{"description":"Technical response identifier equal to the inbound X-Request-Id.","schema":{"type":"string"}}}}},"security":[{"mcpGatewayBearer":[]}],"x-AI-ready":true}},"/api/bc-02/gateway/v1/webhooks/{pluginShortId}":{"post":{"tags":["webhook-dispatcher"],"summary":"Dispatch webhook to plugin pod","description":"Forwards CloudEvents 1.0 payload from core BC to the target plugin pod. Dispatch is async (returns 202 immediately). Retries with exponential backoff.","operationId":"dispatchWebhook","parameters":[{"name":"pluginShortId","in":"path","description":"Plugin short identifier","required":true,"schema":{"type":"string","default":""},"example":"example-pluginShortId"}],"requestBody":{"content":{"application/json":{"schema":{"type":"string","format":"byte"}}},"required":true},"responses":{"202":{"description":"Webhook accepted for async delivery","headers":{"X-Response-Id":{"description":"Technical response identifier equal to the inbound X-Request-Id.","schema":{"type":"string"}}}},"400":{"description":"Invalid plugin short ID","headers":{"X-Response-Id":{"description":"Technical response identifier equal to the inbound X-Request-Id.","schema":{"type":"string"}}}},"403":{"description":"Invalid internal secret","headers":{"X-Response-Id":{"description":"Technical response identifier equal to the inbound X-Request-Id.","schema":{"type":"string"}}}}},"x-AI-ready":true}},"/api/plugins/ai-course-generator/v1/runs/{runId}/events":{"get":{"tags":["first-party-plugin-event-stream-route-controller"],"operationId":"stream","parameters":[{"name":"runId","in":"path","required":true,"schema":{"type":"string","format":"uuid"},"example":"550e8400-e29b-41d4-a716-446655440000"}],"responses":{"200":{"description":"OK","headers":{"X-Response-Id":{"description":"Technical response identifier equal to the inbound X-Request-Id.","schema":{"type":"string"}}},"content":{"text/event-stream":{"schema":{"$ref":"#/components/schemas/StreamingResponseBody"}}}}},"x-AI-ready":true}},"/api/plugins/{slug}/v1/**":{"get":{"tags":["first-party-plugin-rest-route-controller"],"operationId":"proxy","parameters":[{"name":"slug","in":"path","required":true,"schema":{"type":"string"},"example":"example-slug"}],"responses":{"200":{"description":"OK","headers":{"X-Response-Id":{"description":"Technical response identifier equal to the inbound X-Request-Id.","schema":{"type":"string"}}}}},"x-AI-ready":true},"put":{"tags":["first-party-plugin-rest-route-controller"],"operationId":"proxy_2","parameters":[{"name":"slug","in":"path","required":true,"schema":{"type":"string"},"example":"example-slug"}],"responses":{"200":{"description":"OK","headers":{"X-Response-Id":{"description":"Technical response identifier equal to the inbound X-Request-Id.","schema":{"type":"string"}}}}},"x-AI-ready":true},"post":{"tags":["first-party-plugin-rest-route-controller"],"operationId":"proxy_1","parameters":[{"name":"slug","in":"path","required":true,"schema":{"type":"string"},"example":"example-slug"}],"responses":{"200":{"description":"OK","headers":{"X-Response-Id":{"description":"Technical response identifier equal to the inbound X-Request-Id.","schema":{"type":"string"}}}}},"x-AI-ready":true},"delete":{"tags":["first-party-plugin-rest-route-controller"],"operationId":"proxy_3","parameters":[{"name":"slug","in":"path","required":true,"schema":{"type":"string"},"example":"example-slug"}],"responses":{"200":{"description":"OK","headers":{"X-Response-Id":{"description":"Technical response identifier equal to the inbound X-Request-Id.","schema":{"type":"string"}}}}},"x-AI-ready":true},"options":{"tags":["first-party-plugin-rest-route-controller"],"operationId":"proxy_6","parameters":[{"name":"slug","in":"path","required":true,"schema":{"type":"string"},"example":"example-slug"}],"responses":{"200":{"description":"OK","headers":{"X-Response-Id":{"description":"Technical response identifier equal to the inbound X-Request-Id.","schema":{"type":"string"}}}}},"x-AI-ready":true},"head":{"tags":["first-party-plugin-rest-route-controller"],"operationId":"proxy_5","parameters":[{"name":"slug","in":"path","required":true,"schema":{"type":"string"},"example":"example-slug"}],"responses":{"200":{"description":"OK","headers":{"X-Response-Id":{"description":"Technical response identifier equal to the inbound X-Request-Id.","schema":{"type":"string"}}}}},"x-AI-ready":true},"patch":{"tags":["first-party-plugin-rest-route-controller"],"operationId":"proxy_4","parameters":[{"name":"slug","in":"path","required":true,"schema":{"type":"string"},"example":"example-slug"}],"responses":{"200":{"description":"OK","headers":{"X-Response-Id":{"description":"Technical response identifier equal to the inbound X-Request-Id.","schema":{"type":"string"}}}}},"x-AI-ready":true}},"/api/bc-02/gateway/v1/api/**":{"get":{"tags":["plugin-route-resolver"],"summary":"Proxy REST request to target BC service","description":"Resolves the target BC from the URL path and proxies the request with circuit breaker protection. Strips /api/bc-02/gateway/v1 prefix.","operationId":"proxy_7","responses":{"200":{"description":"Response proxied from target BC","headers":{"X-Response-Id":{"description":"Technical response identifier equal to the inbound X-Request-Id.","schema":{"type":"string"}}}},"404":{"description":"Route not configured for the target BC","headers":{"X-Response-Id":{"description":"Technical response identifier equal to the inbound X-Request-Id.","schema":{"type":"string"}}}},"502":{"description":"Target BC returned an error","headers":{"X-Response-Id":{"description":"Technical response identifier equal to the inbound X-Request-Id.","schema":{"type":"string"}}}},"503":{"description":"Circuit breaker open — target BC unavailable","headers":{"X-Response-Id":{"description":"Technical response identifier equal to the inbound X-Request-Id.","schema":{"type":"string"}}}}},"x-AI-ready":true},"put":{"tags":["plugin-route-resolver"],"summary":"Proxy REST request to target BC service","description":"Resolves the target BC from the URL path and proxies the request with circuit breaker protection. Strips /api/bc-02/gateway/v1 prefix.","operationId":"proxy_9","responses":{"200":{"description":"Response proxied from target BC","headers":{"X-Response-Id":{"description":"Technical response identifier equal to the inbound X-Request-Id.","schema":{"type":"string"}}}},"404":{"description":"Route not configured for the target BC","headers":{"X-Response-Id":{"description":"Technical response identifier equal to the inbound X-Request-Id.","schema":{"type":"string"}}}},"502":{"description":"Target BC returned an error","headers":{"X-Response-Id":{"description":"Technical response identifier equal to the inbound X-Request-Id.","schema":{"type":"string"}}}},"503":{"description":"Circuit breaker open — target BC unavailable","headers":{"X-Response-Id":{"description":"Technical response identifier equal to the inbound X-Request-Id.","schema":{"type":"string"}}}}},"x-AI-ready":true},"post":{"tags":["plugin-route-resolver"],"summary":"Proxy REST request to target BC service","description":"Resolves the target BC from the URL path and proxies the request with circuit breaker protection. Strips /api/bc-02/gateway/v1 prefix.","operationId":"proxy_8","responses":{"200":{"description":"Response proxied from target BC","headers":{"X-Response-Id":{"description":"Technical response identifier equal to the inbound X-Request-Id.","schema":{"type":"string"}}}},"404":{"description":"Route not configured for the target BC","headers":{"X-Response-Id":{"description":"Technical response identifier equal to the inbound X-Request-Id.","schema":{"type":"string"}}}},"502":{"description":"Target BC returned an error","headers":{"X-Response-Id":{"description":"Technical response identifier equal to the inbound X-Request-Id.","schema":{"type":"string"}}}},"503":{"description":"Circuit breaker open — target BC unavailable","headers":{"X-Response-Id":{"description":"Technical response identifier equal to the inbound X-Request-Id.","schema":{"type":"string"}}}}},"x-AI-ready":true},"delete":{"tags":["plugin-route-resolver"],"summary":"Proxy REST request to target BC service","description":"Resolves the target BC from the URL path and proxies the request with circuit breaker protection. Strips /api/bc-02/gateway/v1 prefix.","operationId":"proxy_10","responses":{"200":{"description":"Response proxied from target BC","headers":{"X-Response-Id":{"description":"Technical response identifier equal to the inbound X-Request-Id.","schema":{"type":"string"}}}},"404":{"description":"Route not configured for the target BC","headers":{"X-Response-Id":{"description":"Technical response identifier equal to the inbound X-Request-Id.","schema":{"type":"string"}}}},"502":{"description":"Target BC returned an error","headers":{"X-Response-Id":{"description":"Technical response identifier equal to the inbound X-Request-Id.","schema":{"type":"string"}}}},"503":{"description":"Circuit breaker open — target BC unavailable","headers":{"X-Response-Id":{"description":"Technical response identifier equal to the inbound X-Request-Id.","schema":{"type":"string"}}}}},"x-AI-ready":true},"options":{"tags":["plugin-route-resolver"],"summary":"Proxy REST request to target BC service","description":"Resolves the target BC from the URL path and proxies the request with circuit breaker protection. Strips /api/bc-02/gateway/v1 prefix.","operationId":"proxy_13","responses":{"200":{"description":"Response proxied from target BC","headers":{"X-Response-Id":{"description":"Technical response identifier equal to the inbound X-Request-Id.","schema":{"type":"string"}}}},"404":{"description":"Route not configured for the target BC","headers":{"X-Response-Id":{"description":"Technical response identifier equal to the inbound X-Request-Id.","schema":{"type":"string"}}}},"502":{"description":"Target BC returned an error","headers":{"X-Response-Id":{"description":"Technical response identifier equal to the inbound X-Request-Id.","schema":{"type":"string"}}}},"503":{"description":"Circuit breaker open — target BC unavailable","headers":{"X-Response-Id":{"description":"Technical response identifier equal to the inbound X-Request-Id.","schema":{"type":"string"}}}}},"x-AI-ready":true},"head":{"tags":["plugin-route-resolver"],"summary":"Proxy REST request to target BC service","description":"Resolves the target BC from the URL path and proxies the request with circuit breaker protection. Strips /api/bc-02/gateway/v1 prefix.","operationId":"proxy_12","responses":{"200":{"description":"Response proxied from target BC","headers":{"X-Response-Id":{"description":"Technical response identifier equal to the inbound X-Request-Id.","schema":{"type":"string"}}}},"404":{"description":"Route not configured for the target BC","headers":{"X-Response-Id":{"description":"Technical response identifier equal to the inbound X-Request-Id.","schema":{"type":"string"}}}},"502":{"description":"Target BC returned an error","headers":{"X-Response-Id":{"description":"Technical response identifier equal to the inbound X-Request-Id.","schema":{"type":"string"}}}},"503":{"description":"Circuit breaker open — target BC unavailable","headers":{"X-Response-Id":{"description":"Technical response identifier equal to the inbound X-Request-Id.","schema":{"type":"string"}}}}},"x-AI-ready":true},"patch":{"tags":["plugin-route-resolver"],"summary":"Proxy REST request to target BC service","description":"Resolves the target BC from the URL path and proxies the request with circuit breaker protection. Strips /api/bc-02/gateway/v1 prefix.","operationId":"proxy_11","responses":{"200":{"description":"Response proxied from target BC","headers":{"X-Response-Id":{"description":"Technical response identifier equal to the inbound X-Request-Id.","schema":{"type":"string"}}}},"404":{"description":"Route not configured for the target BC","headers":{"X-Response-Id":{"description":"Technical response identifier equal to the inbound X-Request-Id.","schema":{"type":"string"}}}},"502":{"description":"Target BC returned an error","headers":{"X-Response-Id":{"description":"Technical response identifier equal to the inbound X-Request-Id.","schema":{"type":"string"}}}},"503":{"description":"Circuit breaker open — target BC unavailable","headers":{"X-Response-Id":{"description":"Technical response identifier equal to the inbound X-Request-Id.","schema":{"type":"string"}}}}},"x-AI-ready":true}}},"components":{"schemas":{"McpGrantRegistrationRequest":{"type":"object","additionalProperties":false,"default":null,"properties":{"providerBindings":{"type":"array","contains":{"default":""},"description":"Complete unique capability binding set; partial registration is rejected","items":{"$ref":"#/components/schemas/ProviderBindingGrantRequest","default":""},"maxItems":32,"minItems":1,"unevaluatedItems":{"default":""}}},"required":["providerBindings"]},"ProviderBindingGrantRequest":{"type":"object","additionalProperties":false,"default":null,"properties":{"capability":{"type":"string","default":"","description":"Canonical dotted capability selected by Agent Runtime","example":"knowledge.source.search","maxLength":200,"minLength":3,"pattern":"^[a-z][a-z0-9]*(?:-[a-z0-9]+)*(?:\\.[a-z][a-z0-9]*(?:-[a-z0-9]+)*)+$"},"routeSlug":{"type":"string","default":"","description":"Internal Plugin Gateway route slug; never model-visible","example":"ai-methodist","maxLength":128,"minLength":1,"pattern":"^[a-z][a-z0-9]*(?:-[a-z0-9]+)*$"},"toolName":{"type":"string","default":"","description":"Exact MCP provider tool name declared by Lifecycle","example":"search_methodology_rag","maxLength":128,"minLength":1,"pattern":"^[a-z][a-z0-9_-]{0,127}$"},"toolVersion":{"type":"string","default":"","description":"Provider tool contract version","example":"2026.07.1","maxLength":64,"minLength":1,"pattern":"^[A-Za-z0-9][A-Za-z0-9._+-]{0,63}$"},"inputSchemaVersion":{"type":"string","default":"","description":"Provider input-schema version","example":"v1","maxLength":64,"minLength":1,"pattern":"^[A-Za-z0-9][A-Za-z0-9._-]{0,63}$"},"inputSchemaDigest":{"type":"string","default":"","description":"SHA-256 digest of the canonical compact sorted UTF-8 MCP inputSchema JSON; must exactly match Lifecycle and the live provider contract","example":"sha256:0123456789abcdef0123456789abcdef0123456789abcdef0123456789abcdef","pattern":"^sha256:[0-9a-f]{64}$"},"authPolicy":{"type":"string","default":"","description":"Only supported provider authentication policy","enum":["tenant-service-jwt"]},"trustLevel":{"type":"string","default":"","description":"Lifecycle-assigned provider trust tier","enum":["CORE","VERIFIED","COMMUNITY"]},"tenantVisibility":{"type":"string","default":"","description":"Provider data visibility boundary","enum":["tenant"]},"status":{"type":"string","default":"","description":"Only callable provider state","enum":["available"]},"lastValidatedAt":{"type":"string","format":"date-time","default":"","description":"Lifecycle validation instant used for stale-binding detection"},"description":{"type":"string","default":"","description":"Exact reviewed tool description; must match Lifecycle and live MCP metadata","maxLength":4096,"minLength":20,"pattern":"^[^\\u0000-\\u0008\\u000B\\u000C\\u000E-\\u001F\\u007F]*$"}},"required":["authPolicy","capability","description","inputSchemaDigest","inputSchemaVersion","lastValidatedAt","routeSlug","status","tenantVisibility","toolName","toolVersion","trustLevel"]},"StreamingResponseBody":{}},"securitySchemes":{"mcpGatewayBearer":{"type":"http","description":"Short-lived Gateway-audience token issued by OAuth token exchange","in":"header","scheme":"bearer","bearerFormat":"JWT"},"bearerAuth":{"type":"http","description":"JWT bearer token issued by the Adapstory identity provider.","scheme":"bearer","bearerFormat":"JWT"}}}}